requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.2
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
celery==5.4.0
redis==5.0.8
httpx==0.27.2
cryptography==43.0.1
python-jobspy==1.1.82
pandas==2.2.3
pytest==8.3.3
